initial move of Primer::Counter to Primer::Beta::Counter
