+++ docs/src/@primer/gatsby-theme-doctocat/nav.yml
@@ -42,7 +42,7 @@
   - title: CloseButton
     url: "/components/closebutton"
   - title: Counter
-    url: "/components/counter"
+    url: "/components/beta/counter"
   - title: Details
     url: "/components/details"
   - title: Dropdown
