+++ static/audited_at.json
@@ -22,6 +22,7 @@
   "Primer::Beta::BorderBox::Header": "",
   "Primer::Beta::Breadcrumbs": "",
   "Primer::Beta::Breadcrumbs::Item": "",
+  "Primer::Beta::Counter": "",
   "Primer::Beta::Flash": "",
   "Primer::Beta::Text": "",
   "Primer::Beta::Truncate": "",
