+++ static/statuses.json
@@ -22,6 +22,7 @@
   "Primer::Beta::BorderBox::Header": "alpha",
   "Primer::Beta::Breadcrumbs": "beta",
   "Primer::Beta::Breadcrumbs::Item": "alpha",
+  "Primer::Beta::Counter": "beta",
   "Primer::Beta::Flash": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::Beta::Truncate": "beta",
@@ -35,7 +36,7 @@
   "Primer::CloseButton": "beta",
   "Primer::ConditionalWrapper": "alpha",
   "Primer::Content": "stable",
-  "Primer::CounterComponent": "beta",
+  "Primer::CounterComponent": "deprecated",
   "Primer::DetailsComponent": "beta",
   "Primer::Dropdown": "alpha",
   "Primer::Dropdown::Menu": "alpha",
