+++ stories/primer/beta/counter_stories.rb
@@ -1,12 +1,14 @@
 # frozen_string_literal: true
 
-class Primer::CounterComponentStories < ViewComponent::Storybook::Stories
+require "primer/beta/counter"
+
+class Primer::Beta::CounterStories < ViewComponent::Storybook::Stories
   layout "storybook_preview"
 
   story(:counter) do
     controls do
       count 0
-      select(:scheme, Primer::CounterComponent::SCHEME_MAPPINGS.keys, :primary)
+      select(:scheme, Primer::Beta::Counter::SCHEME_MAPPINGS.keys, :primary)
       limit 5000
       hide_if_zero false
       round false
