+++ test/previews/primer/counter_component_preview.rb
@@ -11,12 +11,12 @@ module Primer
     # @param round toggle
     # @param scheme [Symbol] select [[Default, default], [Primary, primary], [Secondary, secondary]]]
     def default(count: 1_000, limit: nil, round: false, hide_if_zero: false, scheme: :default)
-      render(Primer::CounterComponent.new(count: count, round: round, limit: limit, hide_if_zero: hide_if_zero, scheme: scheme))
+      render(Primer::Beta::Counter.new(count: count, round: round, limit: limit, hide_if_zero: hide_if_zero, scheme: scheme))
     end
 
     # @label With Text
     def with_text
-      render(Primer::CounterComponent.new(text: "∞"))
+      render(Primer::Beta::Counter.new(text: "∞"))
     end
   end
 end
